overlays: use the on-board 24MHz oscillator for mclk
The previous overlays required modifying the wm8960 source file as ReSpeaker did. The main effect of its modification is to force the use of SYSCLK_PLL as the clock source. After the modification, you can directly use MCLK  as the clock source and the wm8960 does not need any modification.
https://github.com/respeaker/seeed-voicecard/blob/master/wm8960.c#L1207-L1225